pyproject.toml
[project]
name = "spliit-telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0.0",
    "httpx==0.27.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=20.0",
    "spliit",
]

[tool.uv.sources]
spliit = { git = "https://github.com/guysoft/SpliitApi.git" }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff",
    "ty",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.ty.environment]
python-version = "3.12"

[tool.pytest.ini_options]
markers = [
    "llm: tests that call the real LLM (deselect with -m 'not llm')",
]
